Test unit for Base price component
